added calendar selection validation to dashboardEvents utility
